+++ src/main/java/fr/maxlego08/essentials/commands/commands/essentials/CommandEssentialsConvert.java
@@ -7,6 +7,7 @@ import fr.maxlego08.essentials.api.convert.Convert;
 import fr.maxlego08.essentials.convert.cmi.CMIConvert;
 import fr.maxlego08.essentials.convert.coinsengine.CoinsEngineConvert;
 import fr.maxlego08.essentials.convert.essentialsx.EssentialsXConvert;
+import fr.maxlego08.essentials.convert.huskhomes.HuskHomesConvert;
 import fr.maxlego08.essentials.convert.playervaultx.PlayerVaultXConvert;
 import fr.maxlego08.essentials.convert.sunlight.SunlightConvert;
 import fr.maxlego08.essentials.zutils.utils.commands.VCommand;
@@ -47,6 +48,7 @@ public class CommandEssentialsConvert extends VCommand {
         SUNLIGHT(SunlightConvert.class),
         PLAYERVAULTX(PlayerVaultXConvert.class),
         COINSENGINE(CoinsEngineConvert.class),
+        HUSKHOMES(HuskHomesConvert.class),
 
         ;