+++ src/main/java/fr/maxlego08/essentials/convert/huskhomes/HuskPosition.java
@@ -1,0 +1,10 @@
+package fr.maxlego08.essentials.convert.huskhomes;
+
+public record HuskPosition(int id,
+                           double x,
+                           double y,
+                           double z,
+                           double yaw,
+                           double pitch,
+                           String world_name) {
+}